+++ src/primaite/simulator/network/networks.py
@@ -14,7 +14,7 @@ from primaite.simulator.system.services.dns.dns_client import DNSClient
 from primaite.simulator.system.services.dns.dns_server import DNSServer
 from primaite.simulator.system.services.ftp.ftp_client import FTPClient
 from primaite.simulator.system.services.red_services.data_manipulation_bot import DataManipulationBot
-from primaite.simulator.system.services.web_server.web_server_service import WebServerService
+from primaite.simulator.system.services.web_server.web_server_service import WebServer
 
 
 def client_server_routed() -> Network:
@@ -260,7 +260,7 @@ def arcd_uc2_network() -> Network:
     database_client.run()
     database_client.connect()
 
-    web_server.software_manager.install(WebServerService)
+    web_server.software_manager.install(WebServer)
 
     # register the web_server to a domain
     dns_server_service: DNSServer = domain_controller.software_manager.software["DNSServer"]  # noqa